Face Overwrite Feature
With version 12 of character file, ngs and pso2 face are now separate (pso2 face is stored in alt face).

Trying to load the pso2 face from an older file and only overwrite the pso2 face is not a simple task in-game.

This will allow you to take the model and overwrite current face (and if v12+, overwrite base face)